[WIP]Add driver commands for brcmfmac
***WIP/unstable/incomplete***

This is a temporary solution to use the qcwcn HAL with a mainline brcmfmac driver.

This is required to submit some custom commands  from userspace to the kernel via nl80211.

Requires WIFI_DRIVER_BUILT set to brcmfmac.

Tested with Country Code / 5GHz wifi in Germany --> search for "set country code to: DE" in log.
Change-Id: If203297581d36e6a21260d4da92a8b1d8896243d